+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.9.24](https://search.maven.org/artifact/de.quantummaid/documaid/0.9.27/jar) - 2020-04-27
+## [0.9.25](https://search.maven.org/artifact/de.quantummaid/documaid/0.9.25/jar) - 2020-05-02
 ### ADDED
-- allow relative hugo links within documentation
+- remove sonar suppression annotations and warnings from code snippets.
+
+## [0.9.24](https://search.maven.org/artifact/de.quantummaid/documaid/0.9.24/jar) - 2020-04-27
+### ADDED
+- allow relative hugo links within documentation.
 
 ## [0.9.23](https://search.maven.org/artifact/de.quantummaid/documaid/0.9.23/jar) - 2020-04-21
 ### CHANGED